+++ webapi/src/main/scala/org/knora/webapi/responders/v2/ontology/OntologyHelpers.scala
@@ -21,21 +21,13 @@ import org.knora.webapi.messages.SmartIri
 import org.knora.webapi.messages.StringFormatter
 import org.knora.webapi.messages.ValuesValidator
 import org.knora.webapi.messages.store.triplestoremessages.*
-import org.knora.webapi.messages.twirl.queries.sparql
-import org.knora.webapi.messages.util.ErrorHandlingMap
 import org.knora.webapi.messages.util.rdf.SparqlSelectResult
 import org.knora.webapi.messages.util.rdf.VariableResultsRow
 import org.knora.webapi.messages.v2.responder.ontologymessages.*
 import org.knora.webapi.messages.v2.responder.ontologymessages.OwlCardinality.*
 import org.knora.webapi.messages.v2.responder.standoffmessages.StandoffDataTypeClasses
-import org.knora.webapi.slice.admin.domain.model.User
-import org.knora.webapi.slice.admin.domain.service.KnoraProjectRepo
 import org.knora.webapi.slice.ontology.domain.model.Cardinality.*
 import org.knora.webapi.slice.ontology.repo.model.OntologyCacheData
-import org.knora.webapi.slice.ontology.repo.service.OntologyCache
-import org.knora.webapi.store.triplestore.api.TriplestoreService
-import org.knora.webapi.store.triplestore.api.TriplestoreService.Queries.Construct
-import org.knora.webapi.store.triplestore.api.TriplestoreService.Queries.Select
 
 object OntologyHelpers {
 
@@ -1463,630 +1455,3 @@ object OntologyHelpers {
     )
   }
 }
-
-trait OntologyTriplestoreHelpers {
-
-  /**
-   * Loads a class definition from the triplestore and converts it to a [[ClassInfoContentV2]].
-   *
-   * @param classIri the IRI of the class to be loaded.
-   * @return a [[ClassInfoContentV2]] representing the class definition.
-   */
-  def loadClassDefinition(classIri: SmartIri): Task[ClassInfoContentV2]
-
-  /**
-   * Loads a property definition from the triplestore and converts it to a [[PropertyInfoContentV2]].
-   *
-   * @param propertyIri the IRI of the property to be loaded.
-   * @return a [[PropertyInfoContentV2]] representing the property definition.
-   */
-  def loadPropertyDefinition(propertyIri: SmartIri): Task[PropertyInfoContentV2]
-
-  /**
-   * Reads an ontology's metadata.
-   *
-   * @param internalOntologyIri the ontology's internal IRI.
-   * @return an [[OntologyMetadataV2]], or [[None]] if the ontology is not found.
-   */
-  def loadOntologyMetadata(internalOntologyIri: SmartIri): Task[Option[OntologyMetadataV2]]
-
-  /**
-   * Checks that the last modification date of an ontology is the same as the one we expect it to be. If not, return
-   * an error message fitting for the "before update" case.
-   *
-   * @param internalOntologyIri          the internal IRI of the ontology.
-   * @param expectedLastModificationDate the last modification date that should now be attached to the ontology.
-   * @return a failed Future if the expected last modification date is not found.
-   */
-  def checkOntologyLastModificationDateBeforeUpdate(
-    internalOntologyIri: SmartIri,
-    expectedLastModificationDate: Instant,
-  ): Task[Unit]
-
-  /**
-   * Checks that the last modification date of an ontology is the same as the one we expect it to be. If not, return
-   * an error message fitting for the "after update" case.
-   *
-   * @param internalOntologyIri          the internal IRI of the ontology.
-   * @param expectedLastModificationDate the last modification date that should now be attached to the ontology.
-   * @return a failed Future if the expected last modification date is not found.
-   */
-  def checkOntologyLastModificationDateAfterUpdate(
-    internalOntologyIri: SmartIri,
-    expectedLastModificationDate: Instant,
-  ): Task[Unit]
-
-  /**
-   * Gets the set of subjects that refer to an ontology or its entities.
-   *
-   * @param ontology the ontology.
-   * @return the set of subjects that refer to the ontology or its entities.
-   */
-  def getSubjectsUsingOntology(ontology: ReadOntologyV2): Task[Set[IRI]]
-}
-
-final case class OntologyTriplestoreHelpersLive(
-  triplestore: TriplestoreService,
-  stringFormatter: StringFormatter,
-) extends OntologyTriplestoreHelpers {
-
-  override def loadOntologyMetadata(internalOntologyIri: SmartIri): Task[Option[OntologyMetadataV2]] = {
-    for {
-      _ <- ZIO.when(!internalOntologyIri.getOntologySchema.contains(InternalSchema)) {
-             ZIO.fail(AssertionException(s"Expected an internal ontology IRI: $internalOntologyIri"))
-           }
-      getOntologyInfoResponse <- triplestore.query(Construct(sparql.v2.txt.getOntologyInfo(internalOntologyIri)))
-
-      metadata: Option[OntologyMetadataV2] =
-        if (getOntologyInfoResponse.statements.isEmpty) {
-          None
-        } else {
-          getOntologyInfoResponse.statements.get(
-            internalOntologyIri.toString,
-          ) match {
-            case Some(statements: Seq[(IRI, String)]) =>
-              val statementMap: Map[IRI, Seq[String]] = statements.groupBy { case (pred, _) =>
-                pred
-              }.map { case (pred, predStatements) =>
-                pred -> predStatements.map { case (_, obj) =>
-                  obj
-                }
-              }
-
-              val projectIris: Seq[String] = statementMap.getOrElse(
-                OntologyConstants.KnoraBase.AttachedToProject,
-                throw InconsistentRepositoryDataException(
-                  s"Ontology $internalOntologyIri has no knora-base:attachedToProject",
-                ),
-              )
-              val labels: Seq[String] = statementMap.getOrElse(
-                OntologyConstants.Rdfs.Label,
-                Seq.empty[String],
-              )
-              val comments: Seq[String] = statementMap.getOrElse(
-                OntologyConstants.Rdfs.Comment,
-                Seq.empty[String],
-              )
-              val lastModDates: Seq[String] =
-                statementMap.getOrElse(
-                  OntologyConstants.KnoraBase.LastModificationDate,
-                  Seq.empty[String],
-                )
-
-              val projectIri = if (projectIris.size > 1) {
-                throw InconsistentRepositoryDataException(
-                  s"Ontology $internalOntologyIri has more than one knora-base:attachedToProject",
-                )
-              } else {
-                stringFormatter.toSmartIri(projectIris.head)
-              }
-
-              if (!internalOntologyIri.isKnoraBuiltInDefinitionIri) {
-                if (projectIri.toString == KnoraProjectRepo.builtIn.SystemProject.id.value) {
-                  throw InconsistentRepositoryDataException(
-                    s"Ontology $internalOntologyIri cannot be in project ${KnoraProjectRepo.builtIn.SystemProject.id.value}",
-                  )
-                }
-
-                if (
-                  internalOntologyIri.isKnoraSharedDefinitionIri && projectIri.toString != OntologyConstants.KnoraAdmin.DefaultSharedOntologiesProject
-                ) {
-                  throw InconsistentRepositoryDataException(
-                    s"Shared ontology $internalOntologyIri must be in project ${OntologyConstants.KnoraAdmin.DefaultSharedOntologiesProject}",
-                  )
-                }
-              }
-
-              val label: String = if (labels.size > 1) {
-                throw InconsistentRepositoryDataException(
-                  s"Ontology $internalOntologyIri has more than one rdfs:label",
-                )
-              } else if (labels.isEmpty) {
-                internalOntologyIri.getOntologyName
-              } else {
-                labels.head
-              }
-
-              val comment: Option[String] = if (comments.size > 1) {
-                throw InconsistentRepositoryDataException(
-                  s"Ontology $internalOntologyIri has more than one rdfs:comment",
-                )
-              } else comments.headOption
-
-              val lastModificationDate: Option[Instant] =
-                if (lastModDates.size > 1) {
-                  throw InconsistentRepositoryDataException(
-                    s"Ontology $internalOntologyIri has more than one ${OntologyConstants.KnoraBase.LastModificationDate}",
-                  )
-                } else if (lastModDates.isEmpty) {
-                  None
-                } else {
-                  val dateStr = lastModDates.head
-                  Some(
-                    ValuesValidator
-                      .xsdDateTimeStampToInstant(dateStr)
-                      .getOrElse(
-                        throw InconsistentRepositoryDataException(
-                          s"Invalid ${OntologyConstants.KnoraBase.LastModificationDate}: $dateStr",
-                        ),
-                      ),
-                  )
-                }
-
-              Some(
-                OntologyMetadataV2(
-                  ontologyIri = internalOntologyIri,
-                  projectIri = Some(projectIri),
-                  label = Some(label),
-                  comment = comment,
-                  lastModificationDate = lastModificationDate,
-                ),
-              )
-
-            case None => None
-          }
-        }
-    } yield metadata
-  }
-
-  /**
-   * Checks that the last modification date of an ontology is the same as the one we expect it to be.
-   *
-   * @param internalOntologyIri          the internal IRI of the ontology.
-   * @param expectedLastModificationDate the last modification date that the ontology is expected to have.
-   * @param errorFun                     a function that throws an exception. It will be called if the expected last modification date is not found.
-   * @return a failed Future if the expected last modification date is not found.
-   */
-  private def checkOntologyLastModificationDate(
-    internalOntologyIri: SmartIri,
-    expectedLastModificationDate: Instant,
-    errorFun: => Nothing,
-  ): Task[Unit] =
-    for {
-      existingOntologyMetadata <- loadOntologyMetadata(internalOntologyIri)
-
-      _ = existingOntologyMetadata match {
-            case Some(metadata) =>
-              metadata.lastModificationDate match {
-                case Some(lastModificationDate) =>
-                  if (lastModificationDate != expectedLastModificationDate) {
-                    errorFun
-                  }
-
-                case None =>
-                  throw InconsistentRepositoryDataException(
-                    s"Ontology $internalOntologyIri has no ${OntologyConstants.KnoraBase.LastModificationDate}",
-                  )
-              }
-
-            case None =>
-              throw NotFoundException(
-                s"Ontology $internalOntologyIri (corresponding to ${internalOntologyIri.toOntologySchema(ApiV2Complex)}) not found",
-              )
-          }
-    } yield ()
-
-  override def checkOntologyLastModificationDateBeforeUpdate(
-    internalOntologyIri: SmartIri,
-    expectedLastModificationDate: Instant,
-  ): Task[Unit] =
-    checkOntologyLastModificationDate(
-      internalOntologyIri = internalOntologyIri,
-      expectedLastModificationDate = expectedLastModificationDate,
-      errorFun = throw EditConflictException(
-        s"Ontology ${internalOntologyIri.toOntologySchema(ApiV2Complex)} has been modified by another user, please reload it and try again.",
-      ),
-    )
-
-  override def checkOntologyLastModificationDateAfterUpdate(
-    internalOntologyIri: SmartIri,
-    expectedLastModificationDate: Instant,
-  ): Task[Unit] =
-    checkOntologyLastModificationDate(
-      internalOntologyIri = internalOntologyIri,
-      expectedLastModificationDate = expectedLastModificationDate,
-      errorFun = throw UpdateNotPerformedException(
-        s"Ontology ${internalOntologyIri.toOntologySchema(ApiV2Complex)} was not updated. Please report this as a possible bug.",
-      ),
-    )
-
-  override def getSubjectsUsingOntology(ontology: ReadOntologyV2): Task[Set[IRI]] =
-    for {
-      isOntologyUsedSparql <- ZIO.attempt(
-                                sparql.v2.txt
-                                  .isOntologyUsed(
-                                    ontologyNamedGraphIri = ontology.ontologyMetadata.ontologyIri,
-                                    classIris = ontology.classes.keySet,
-                                    propertyIris = ontology.properties.keySet,
-                                  ),
-                              )
-
-      isOntologyUsedResponse <- triplestore.query(Select(isOntologyUsedSparql))
-
-      subjects = isOntologyUsedResponse.getColOrThrow("s").toSet
-    } yield subjects
-
-  override def loadPropertyDefinition(propertyIri: SmartIri): Task[PropertyInfoContentV2] =
-    triplestore
-      .query(Construct(sparql.v2.txt.getPropertyDefinition(propertyIri)))
-      .flatMap(_.asExtended(stringFormatter))
-      .map(constructResponse =>
-        OntologyHelpers.constructResponseToPropertyDefinition(propertyIri, constructResponse)(stringFormatter),
-      )
-
-  override def loadClassDefinition(classIri: SmartIri): Task[ClassInfoContentV2] =
-    triplestore
-      .query(Construct(sparql.v2.txt.getClassDefinition(classIri)))
-      .flatMap(_.asExtended(stringFormatter))
-      .map(OntologyHelpers.constructResponseToClassDefinition(classIri, _)(stringFormatter))
-}
-
-object OntologyTriplestoreHelpersLive { val layer = ZLayer.derive[OntologyTriplestoreHelpersLive] }
-
-trait OntologyCacheHelpers {
-
-  /**
-   * Requests information about OWL classes in a single ontology.
-   *
-   * @param classIris      the IRIs (internal or external) of the classes to query for.
-   * @param requestingUser the user making the request.
-   * @return a [[ReadOntologyV2]].
-   */
-  def getClassDefinitionsFromOntologyV2(
-    classIris: Set[SmartIri],
-    allLanguages: Boolean,
-    requestingUser: User,
-  ): Task[ReadOntologyV2]
-
-  /**
-   * Given a list of resource IRIs and a list of property IRIs (ontology entities), returns an [[EntityInfoGetResponseV2]] describing both resource and property entities.
-   *
-   * @param classIris      the IRIs of the resource entities to be queried.
-   * @param propertyIris   the IRIs of the property entities to be queried.
-   * @param requestingUser the user making the request.
-   * @return an [[EntityInfoGetResponseV2]].
-   */
-  def getEntityInfoResponseV2(
-    classIris: Set[SmartIri] = Set.empty[SmartIri],
-    propertyIris: Set[SmartIri] = Set.empty[SmartIri],
-    requestingUser: User,
-  ): Task[EntityInfoGetResponseV2]
-
-  /**
-   * Before an update of an ontology entity, checks that the entity's external IRI, and that of its ontology,
-   * are valid, and checks that the user has permission to update the ontology.
-   *
-   * @param externalOntologyIri the external IRI of the ontology.
-   * @param externalEntityIri   the external IRI of the entity.
-   * @param requestingUser      the user making the request.
-   */
-  final def checkOntologyAndEntityIrisForUpdate(
-    externalOntologyIri: SmartIri,
-    externalEntityIri: SmartIri,
-    requestingUser: User,
-  ): Task[Unit] =
-    for {
-      _ <- OntologyHelpers.checkExternalOntologyIriForUpdate(externalOntologyIri)
-      _ <- OntologyHelpers.checkExternalEntityIriForUpdate(externalEntityIri)
-      _ <- checkPermissionsForOntologyUpdate(
-             internalOntologyIri = externalOntologyIri.toOntologySchema(InternalSchema),
-             requestingUser = requestingUser,
-           )
-    } yield ()
-
-  /**
-   * Throws an exception if the requesting user does not have permission to update an ontology.
-   *
-   * @param internalOntologyIri the internal IRI of the ontology.
-   * @param requestingUser      the user making the request.
-   * @return the project IRI.
-   */
-  def checkPermissionsForOntologyUpdate(internalOntologyIri: SmartIri, requestingUser: User): Task[SmartIri]
-
-  /**
-   * Checks whether the requesting user has permission to update an ontology.
-   *
-   * @param internalOntologyIri the internal IRI of the ontology.
-   * @param requestingUser      the user making the request.
-   * @return `true` if the user has permission to update the ontology
-   */
-  def canUserUpdateOntology(internalOntologyIri: SmartIri, requestingUser: User): Task[Boolean]
-}
-
-final case class OntologyCacheHelpersLive(
-  ontologyCache: OntologyCache,
-) extends OntologyCacheHelpers {
-
-  override def getClassDefinitionsFromOntologyV2(
-    classIris: Set[SmartIri],
-    allLanguages: Boolean,
-    requestingUser: User,
-  ): Task[ReadOntologyV2] =
-    for {
-      cacheData <- ontologyCache.getCacheData
-
-      ontologyIris = classIris.map(_.getOntologyFromEntity)
-
-      _ = if (ontologyIris.size != 1) {
-            throw BadRequestException(s"Only one ontology may be queried per request")
-          }
-
-      classInfoResponse  <- getEntityInfoResponseV2(classIris = classIris, requestingUser = requestingUser)
-      internalOntologyIri = ontologyIris.head.toOntologySchema(InternalSchema)
-
-      // Are we returning data in the user's preferred language, or in all available languages?
-      userLang =
-        if (!allLanguages) {
-          // Just the user's preferred language.
-          Some(requestingUser.lang)
-        } else {
-          // All available languages.
-          None
-        }
-    } yield ReadOntologyV2(
-      ontologyMetadata = cacheData.ontologies(internalOntologyIri).ontologyMetadata,
-      classes = classInfoResponse.classInfoMap,
-      userLang = userLang,
-    )
-
-  override def getEntityInfoResponseV2(
-    classIris: Set[SmartIri] = Set.empty[SmartIri],
-    propertyIris: Set[SmartIri] = Set.empty[SmartIri],
-    requestingUser: User,
-  ): Task[EntityInfoGetResponseV2] = {
-    for {
-      cacheData <- ontologyCache.getCacheData
-
-      // See if any of the requested entities are not Knora entities.
-
-      nonKnoraEntities = (classIris ++ propertyIris).filter(!_.isKnoraEntityIri)
-
-      _ = if (nonKnoraEntities.nonEmpty) {
-            throw BadRequestException(
-              s"Some requested entities are not Knora entities: ${nonKnoraEntities.mkString(", ")}",
-            )
-          }
-
-      // See if any of the requested entities are unavailable in the requested schema.
-
-      classesUnavailableInSchema = classIris.foldLeft(Set.empty[SmartIri]) { case (acc, classIri) =>
-                                     // Is this class IRI hard-coded in the requested schema?
-                                     if (
-                                       KnoraBaseToApiV2SimpleTransformationRules.externalClassesToAdd
-                                         .contains(classIri) ||
-                                       KnoraBaseToApiV2ComplexTransformationRules.externalClassesToAdd
-                                         .contains(classIri)
-                                     ) {
-                                       // Yes, so it's available.
-                                       acc
-                                     } else {
-                                       // No. Is it among the classes removed from the internal ontology in the requested schema?
-                                       classIri.getOntologySchema.get match {
-                                         case apiV2Schema: ApiV2Schema =>
-                                           val internalClassIri =
-                                             classIri.toOntologySchema(InternalSchema)
-                                           val knoraBaseClassesToRemove = OntologyTransformationRules
-                                             .getTransformationRules(
-                                               apiV2Schema,
-                                             )
-                                             .internalClassesToRemove
-
-                                           if (knoraBaseClassesToRemove.contains(internalClassIri)) {
-                                             // Yes. Include it in the set of unavailable classes.
-                                             acc + classIri
-                                           } else {
-                                             // No. It's available.
-                                             acc
-                                           }
-
-                                         case InternalSchema => acc
-                                       }
-                                     }
-                                   }
-
-      propertiesUnavailableInSchema = propertyIris.foldLeft(Set.empty[SmartIri]) { case (acc, propertyIri) =>
-                                        // Is this property IRI hard-coded in the requested schema?
-                                        if (
-                                          KnoraBaseToApiV2SimpleTransformationRules.externalPropertiesToAdd
-                                            .contains(propertyIri) ||
-                                          KnoraBaseToApiV2ComplexTransformationRules.externalPropertiesToAdd
-                                            .contains(propertyIri)
-                                        ) {
-                                          // Yes, so it's available.
-                                          acc
-                                        } else {
-                                          // No. See if it's available in the requested schema.
-                                          propertyIri.getOntologySchema.get match {
-                                            case apiV2Schema: ApiV2Schema =>
-                                              val internalPropertyIri =
-                                                propertyIri.toOntologySchema(InternalSchema)
-
-                                              // If it's a link value property and it's requested in the simple schema, it's unavailable.
-                                              if (
-                                                apiV2Schema == ApiV2Simple && OntologyHelpers
-                                                  .isLinkValueProp(
-                                                    internalPropertyIri,
-                                                    cacheData,
-                                                  )
-                                              ) {
-                                                acc + propertyIri
-                                              } else {
-                                                // Is it among the properties removed from the internal ontology in the requested schema?
-
-                                                val knoraBasePropertiesToRemove =
-                                                  OntologyTransformationRules
-                                                    .getTransformationRules(
-                                                      apiV2Schema,
-                                                    )
-                                                    .internalPropertiesToRemove
-
-                                                if (
-                                                  knoraBasePropertiesToRemove.contains(
-                                                    internalPropertyIri,
-                                                  )
-                                                ) {
-                                                  // Yes. Include it in the set of unavailable properties.
-                                                  acc + propertyIri
-                                                } else {
-                                                  // No. It's available.
-                                                  acc
-                                                }
-                                              }
-
-                                            case InternalSchema => acc
-                                          }
-                                        }
-                                      }
-
-      entitiesUnavailableInSchema = classesUnavailableInSchema ++ propertiesUnavailableInSchema
-
-      _ <- ZIO.when(entitiesUnavailableInSchema.nonEmpty) {
-             ZIO.fail(
-               NotFoundException(
-                 s"Some requested entities were not found: ${entitiesUnavailableInSchema.mkString(", ")}",
-               ),
-             )
-           }
-
-      // See if any of the requested entities are hard-coded for knora-api.
-      hardCodedExternalClassesAvailable =
-        KnoraBaseToApiV2SimpleTransformationRules.externalClassesToAdd.view
-          .filterKeys(classIris)
-          .toMap ++
-          KnoraBaseToApiV2ComplexTransformationRules.externalClassesToAdd.view.filterKeys(classIris).toMap
-
-      hardCodedExternalPropertiesAvailable =
-        KnoraBaseToApiV2SimpleTransformationRules.externalPropertiesToAdd.view
-          .filterKeys(propertyIris)
-          .toMap ++
-          KnoraBaseToApiV2ComplexTransformationRules.externalPropertiesToAdd.view.filterKeys(propertyIris)
-
-      // Convert the remaining external entity IRIs to internal ones.
-
-      internalToExternalClassIris =
-        (classIris -- hardCodedExternalClassesAvailable.keySet)
-          .map(externalIri => externalIri.toOntologySchema(InternalSchema) -> externalIri)
-          .toMap
-      internalToExternalPropertyIris =
-        (propertyIris -- hardCodedExternalPropertiesAvailable.keySet)
-          .map(externalIri => externalIri.toOntologySchema(InternalSchema) -> externalIri)
-          .toMap
-
-      classIrisForCache    = internalToExternalClassIris.keySet
-      propertyIrisForCache = internalToExternalPropertyIris.keySet
-
-      // Get the entities that are available in the ontology cache.
-
-      classOntologiesForCache =
-        cacheData.ontologies.view.filterKeys(classIrisForCache.map(_.getOntologyFromEntity)).toMap.values
-
-      propertyOntologiesForCache =
-        cacheData.ontologies.view.filterKeys(propertyIrisForCache.map(_.getOntologyFromEntity)).toMap.values
-
-      classesAvailableFromCache =
-        classOntologiesForCache.flatMap(ontology => ontology.classes.view.filterKeys(classIrisForCache).toMap).toMap
-
-      propertiesAvailableFromCache = propertyOntologiesForCache.flatMap { ontology =>
-                                       ontology.properties.view.filterKeys(propertyIrisForCache).toMap
-                                     }.toMap
-
-      allClassesAvailable    = classesAvailableFromCache ++ hardCodedExternalClassesAvailable
-      allPropertiesAvailable = propertiesAvailableFromCache ++ hardCodedExternalPropertiesAvailable
-
-      // See if any entities are missing.
-
-      allExternalClassIrisAvailable = allClassesAvailable.keySet.map { classIri =>
-                                        if (classIri.getOntologySchema.contains(InternalSchema)) {
-                                          internalToExternalClassIris(classIri)
-                                        } else {
-                                          classIri
-                                        }
-                                      }
-
-      allExternalPropertyIrisAvailable = allPropertiesAvailable.keySet.map { propertyIri =>
-                                           if (propertyIri.getOntologySchema.contains(InternalSchema)) {
-                                             internalToExternalPropertyIris(propertyIri)
-                                           } else {
-                                             propertyIri
-                                           }
-                                         }
-
-      missingClasses    = classIris -- allExternalClassIrisAvailable
-      missingProperties = propertyIris -- allExternalPropertyIrisAvailable
-
-      missingEntities = missingClasses ++ missingProperties
-
-      _ <- ZIO.when(missingEntities.nonEmpty) {
-             ZIO.fail(NotFoundException(s"Some requested entities were not found: ${missingEntities.mkString(", ")}"))
-           }
-
-      response = EntityInfoGetResponseV2(
-                   classInfoMap = new ErrorHandlingMap(allClassesAvailable, key => s"Resource class $key not found"),
-                   propertyInfoMap = new ErrorHandlingMap(allPropertiesAvailable, key => s"Property $key not found"),
-                 )
-    } yield response
-  }
-
-  override def checkPermissionsForOntologyUpdate(
-    internalOntologyIri: SmartIri,
-    requestingUser: User,
-  ): Task[SmartIri] =
-    for {
-      cacheData <- ontologyCache.getCacheData
-
-      projectIri =
-        cacheData.ontologies
-          .getOrElse(
-            internalOntologyIri,
-            throw NotFoundException(s"Ontology ${internalOntologyIri.toOntologySchema(ApiV2Complex)} not found"),
-          )
-          .ontologyMetadata
-          .projectIri
-          .get
-
-      _ = if (
-            !requestingUser.permissions.isProjectAdmin(projectIri.toString) && !requestingUser.permissions.isSystemAdmin
-          ) {
-            // not a project or system admin
-            throw ForbiddenException("Ontologies can be modified only by a project or system admin.")
-          }
-
-    } yield projectIri
-
-  override def canUserUpdateOntology(internalOntologyIri: SmartIri, requestingUser: User): Task[Boolean] =
-    for {
-      cacheData <- ontologyCache.getCacheData
-
-      projectIri =
-        cacheData.ontologies
-          .getOrElse(
-            internalOntologyIri,
-            throw NotFoundException(s"Ontology ${internalOntologyIri.toOntologySchema(ApiV2Complex)} not found"),
-          )
-          .ontologyMetadata
-          .projectIri
-          .get
-    } yield requestingUser.permissions.isProjectAdmin(projectIri.toString) || requestingUser.permissions.isSystemAdmin
-}
-
-object OntologyCacheHelpersLive { val layer = ZLayer.derive[OntologyCacheHelpersLive] }
